MF-15 moved api request to new domain proj.
