+++ src/SFA.DAS.Reservations.Application.UnitTests/Reservations/Commands/WhenCreatingANewReservation.cs
@@ -11,6 +11,7 @@ using NUnit.Framework;
 using SFA.DAS.Reservations.Application.Reservations.Commands;
 using SFA.DAS.Reservations.Application.Reservations.Services;
 using SFA.DAS.Reservations.Application.Validation;
+using SFA.DAS.Reservations.Domain.ReservationsApi;
 using SFA.DAS.Reservations.Infrastructure.Api;
 using SFA.DAS.Reservations.Models;
