+++ src/SFA.DAS.Reservations.Application/Reservations/Commands/CreateReservationCommandHandler.cs
@@ -6,6 +6,7 @@ using MediatR;
 using Newtonsoft.Json;
 using SFA.DAS.Reservations.Application.Reservations.Services;
 using SFA.DAS.Reservations.Application.Validation;
+using SFA.DAS.Reservations.Domain.ReservationsApi;
 using SFA.DAS.Reservations.Infrastructure.Api;
 using SFA.DAS.Reservations.Models;
