+++ src/SFA.DAS.Reservations.Domain/ReservationsApi/CreateReservationApiRequest.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace SFA.DAS.Reservations.Application.Reservations.Commands
+namespace SFA.DAS.Reservations.Domain.ReservationsApi
 {
     public class CreateReservationApiRequest
     {
