+++ src/SFA.DAS.Reservations.Web.sln
@@ -13,11 +13,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{429FE184
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.Reservations.Infrastructure.UnitTests", "SFA.DAS.Reservations.Infrastructure.UnitTests\SFA.DAS.Reservations.Infrastructure.UnitTests.csproj", "{DDDA5440-96C8-452D-A4FD-FF09581E2738}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SFA.DAS.Reservations.Application", "SFA.DAS.Reservations.Application\SFA.DAS.Reservations.Application.csproj", "{97D5D578-B8A2-46F9-AA1B-BBED699887D7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.Reservations.Application", "SFA.DAS.Reservations.Application\SFA.DAS.Reservations.Application.csproj", "{97D5D578-B8A2-46F9-AA1B-BBED699887D7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SFA.DAS.Reservations.Application.UnitTests", "SFA.DAS.Reservations.Application.UnitTests\SFA.DAS.Reservations.Application.UnitTests.csproj", "{36E04F1E-9E5A-4A39-AB57-FFCF09EE4C57}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.Reservations.Application.UnitTests", "SFA.DAS.Reservations.Application.UnitTests\SFA.DAS.Reservations.Application.UnitTests.csproj", "{36E04F1E-9E5A-4A39-AB57-FFCF09EE4C57}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SFA.DAS.Reservations.Web.UnitTests", "SFA.DAS.Reservations.Web.UnitTests\SFA.DAS.Reservations.Web.UnitTests.csproj", "{DD46209A-CA75-4223-B61B-F7D355DD010C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.Reservations.Web.UnitTests", "SFA.DAS.Reservations.Web.UnitTests\SFA.DAS.Reservations.Web.UnitTests.csproj", "{DD46209A-CA75-4223-B61B-F7D355DD010C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SFA.DAS.Reservations.Domain", "SFA.DAS.Reservations.Domain\SFA.DAS.Reservations.Domain.csproj", "{945C9EDA-02A4-4652-9BDD-4535714FE430}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -53,6 +55,10 @@ Global
 		{DD46209A-CA75-4223-B61B-F7D355DD010C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DD46209A-CA75-4223-B61B-F7D355DD010C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DD46209A-CA75-4223-B61B-F7D355DD010C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{945C9EDA-02A4-4652-9BDD-4535714FE430}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{945C9EDA-02A4-4652-9BDD-4535714FE430}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{945C9EDA-02A4-4652-9BDD-4535714FE430}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{945C9EDA-02A4-4652-9BDD-4535714FE430}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
